changed test files
