+++ src/tests/win-app-driver/FancyZonesTests/EditorOpeningTests.cs
@@ -1,5 +1,3 @@
-using OpenQA.Selenium;
-﻿using System;
 using System.IO;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OpenQA.Selenium.Appium.Windows;
@@ -7,7 +5,7 @@ using OpenQA.Selenium.Appium.Windows;
 namespace PowerToysTests
 {
     [TestClass]
-    public class FancyZonesEditorTests : PowerToysSession
+    public class FancyZonesEditorOpeningTests : PowerToysSession
     {
         WindowsElement editorWindow;
 
@@ -143,53 +141,6 @@ namespace PowerToysTests
 
             editorWindow = session.FindElementByXPath("//Window[@Name=\"FancyZones Editor\"]");
             Assert.IsNotNull(editorWindow);
-        }
-
-        [TestMethod]
-        public void ZoneCount()
-        {
-
-            WaitSeconds(1);
-
-            OpenFancyZonesSettings();
-
-            WaitSeconds(1);
-
-            WindowsElement editorButton = session.FindElementByXPath("//Button[@Name=\"Edit zones\"]");
-            Assert.IsNotNull(editorButton);
-            editorButton.Click();
-
-            WaitSeconds(1);
-
-            WindowsElement minusButton = session.FindElementByAccessibilityId("decrementZones");
-            Assert.IsNotNull(minusButton);
-
-            WindowsElement plusButton = session.FindElementByAccessibilityId("incrementZones");
-            Assert.IsNotNull(plusButton);
-
-            WindowsElement zoneCount = session.FindElementByAccessibilityId("zoneCount");
-            Assert.IsNotNull(zoneCount);
-
-            int zoneCountQty;
-            Assert.IsTrue(Int32.TryParse(zoneCount.Text, out zoneCountQty));
-
-            for (int i = zoneCountQty; i > -5; --i)
-            {
-                Assert.IsTrue(Int32.TryParse(zoneCount.Text, out zoneCountQty));
-                Assert.AreEqual(Math.Max(i, 1), zoneCountQty);
-                minusButton.Click();
-            }
-
-            for (int i = 1; i < 45; ++i)
-            {
-                Assert.IsTrue(Int32.TryParse(zoneCount.Text, out zoneCountQty));
-                Assert.AreEqual(Math.Min(i, 40), zoneCountQty);
-                plusButton.Click();
-            }
-
-            WindowsElement mainWindow = session.FindElementByAccessibilityId("MainWindow1");
-            Assert.IsNotNull(mainWindow);
-            mainWindow.SendKeys(Keys.Alt + Keys.F4);
         }
 
         [ClassInitialize]
